Implement new device selection logic for multi-GPU
We now try to find multiple eligible devices (peferrably GPUs) for each
worker node. If there are multiple worker nodes on a single host, we
attempt to evenly distribute the devices across them (this means that
"legacy mode" can be emulated by spawning the same number of nodes as
(visible) GPUs on each host).

- All devices must use the same backend.
- We now check that required devices aspects (such as USM) are
  supported.
- Remove distr_queue constructor and runtime::init overloads receiving a
  single SYCL device.
   - Add new overloads that receive a vector of devices.
- Remove CELERITY_DEVICES environment variable.